ggml : move common CPU backend impl to new header (#9509)
